+++ AmbilightRaspicamDMX/olawrapper.h
@@ -1,5 +1,6 @@
 #ifndef OLAWRAPPER_H
 #define OLAWRAPPER_H
+#include <cstdint>
 #include <ola/DmxBuffer.h>
 #include <ola/client/StreamingClient.h>
 
@@ -8,12 +9,12 @@ class OlaWrapper
 {
 public:
     OlaWrapper();
-    void setChannel(unsigned int channel, unsigned int value);
+    void setChannel(uint8_t channel, uint8_t value);
     void blackout();
     void send();
 
 private:
-    unsigned int universe;
+    uint8_t universe;
     ola::DmxBuffer buffer; // A DmxBuffer to hold the data.
     ola::client::StreamingClient ola_client;
 };
